Refactor classes into packages
