+++ src/main/java/bruno/exceptions/BrunoDateTimeFormatException.java
@@ -1,5 +1,7 @@
+package bruno.exceptions;
+
 public class BrunoDateTimeFormatException extends BrunoException {
-    BrunoDateTimeFormatException() {
+    public BrunoDateTimeFormatException() {
         super("Date and Time are not in correct format. They must be in YYYY-MM-DD HH:MM format.");
     }
 }
