+++ src/main/java/bruno/exceptions/BrunoEmptyException.java
@@ -1,7 +1,9 @@
+package bruno.exceptions;
+
 public class BrunoEmptyException extends BrunoException {
     String type;
 
-    BrunoEmptyException(String type) {
+    public BrunoEmptyException(String type) {
         super("Ruff Ruff! Description of " + type + " cannot be empty! ❌");
     }
 }
