+++ src/main/java/bruno/exceptions/BrunoIndexOutOfBoundsException.java
@@ -1,7 +1,9 @@
+package bruno.exceptions;
+
 public class BrunoIndexOutOfBoundsException extends BrunoException {
     String activity;
 
-    BrunoIndexOutOfBoundsException(String activity) {
+    public BrunoIndexOutOfBoundsException(String activity) {
         super("Ruff Ruff! Task numbers to be " + activity
                 + "ed cannot be greater than number of tasks in the list! ❌");
     }
