+++ src/main/java/bruno/exceptions/BrunoIntegerMismatchException.java
@@ -1,7 +1,9 @@
+package bruno.exceptions;
+
 public class BrunoIntegerMismatchException extends BrunoException {
     String activity;
 
-    BrunoIntegerMismatchException(String activity) {
+    public BrunoIntegerMismatchException(String activity) {
         super("Ruff Ruff! Tasks to be " + activity + "ed can only be integers! ❌");
     }
 }
