+++ src/main/java/bruno/exceptions/BrunoMissingDeadlineException.java
@@ -1,5 +1,7 @@
+package bruno.exceptions;
+
 public class BrunoMissingDeadlineException extends BrunoException {
-    BrunoMissingDeadlineException() {
+    public BrunoMissingDeadlineException() {
         super("Ruff Ruff! You cannot add a Deadline task without setting the deadline! ❌");
     }
 }
