+++ src/main/java/bruno/exceptions/BrunoMissingEventException.java
@@ -1,5 +1,7 @@
+package bruno.exceptions;
+
 public class BrunoMissingEventException extends BrunoException {
-    BrunoMissingEventException() {
+    public BrunoMissingEventException() {
         super("Ruff Ruff! You cannot add an Event task without setting start and end time! ❌");
     }
 }
