+++ src/main/java/bruno/exceptions/BrunoNegativeArgException.java
@@ -1,7 +1,9 @@
+package bruno.exceptions;
+
 public class BrunoNegativeArgException extends BrunoException {
     String activity;
 
-    BrunoNegativeArgException(String activity) {
+    public BrunoNegativeArgException(String activity) {
         super("Ruff Ruff! Task numbers to be " + activity + "ed cannot be negative! ❌");
     }
 }
