+++ src/main/java/bruno/exceptions/BrunoUnknownTaskException.java
@@ -1,5 +1,7 @@
+package bruno.exceptions;
+
 public class BrunoUnknownTaskException extends BrunoException {
-    BrunoUnknownTaskException() {
+    public BrunoUnknownTaskException() {
         super("I am sorry, I do not understand what you mean.");
     }
 }
