+++ src/main/java/bruno/task/TaskType.java
@@ -1,3 +1,5 @@
+package bruno.task;
+
 public enum TaskType {
     TODO, DEADLINE, EVENT, MARK, UNMARK, LIST, BYE, DELETE, SCHEDULE
 }
