+++ src/main/java/bruno/task/ToDo.java
@@ -1,3 +1,5 @@
+package bruno.task;
+
 public class ToDo extends Task {
     public ToDo(String description) {
         super(TaskType.TODO, description);
