+++ text-ui-test/EXPECTED.TXT
@@ -5,7 +5,7 @@
      ;   ,~.——;  /
      :  |     :  |
      `_ ’     `_ ‘
-Woof Woof! I'm Bruno 🐾
+Woof Woof! I'm bruno.Bruno 🐾
 How can I help you?
 - - - - - - - - - - - - - - - - - - - -
 - - - - - - - - - - - - - - - - - - - -
